Rename landing page v2 folder
