+++ packages/web/src/pages/landing-page/components/ArtistInvestors.tsx
