+++ packages/web/src/pages/landing-page/components/CTAGetStarted.tsx
