+++ packages/web/src/pages/landing-page/components/CaseStudies.tsx
