+++ packages/web/src/pages/landing-page/components/Description.tsx
