+++ packages/web/src/pages/landing-page/components/FeaturedContent.tsx
