+++ packages/web/src/pages/landing-page/components/Hero.tsx
