+++ packages/web/src/pages/landing-page/components/PlatformFeatures.tsx
